Finalizado os testes dos Controllers e Repositories para compatibilizacao com a geracao do token

+++ src/main/java/br/com/alura/carteira/service/TransacaoService.java
@@ -1,7 +1,5 @@
 package br.com.alura.carteira.service;
 
-
-
 import javax.persistence.EntityNotFoundException;
 
 import org.springframework.transaction.annotation.Transactional;
@@ -10,6 +8,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -21,64 +20,76 @@ import br.com.alura.carteira.dto.TransacaoUpdateFormDto;
 import br.com.alura.carteira.exceptions.DomainException;
 import br.com.alura.carteira.exceptions.ResourceNotFoundException;
 import br.com.alura.carteira.modelo.Transacao;
-
+import br.com.alura.carteira.modelo.Usuario;
 import br.com.alura.carteira.repository.TransacaoRepository;
 import br.com.alura.carteira.repository.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
 
-
 @Service
 @RequiredArgsConstructor
 public class TransacaoService {
-	
+
 	@Autowired
 	private TransacaoRepository transacaoRepository;
 	@Autowired
 	private UsuarioRepository usuarioRepository;
 	@Autowired
 	private ModelMapper modelMapper;
-	
+
 	@Transactional(readOnly = true)
-	public Page<TransacaoDto> getTransacoes(Pageable paginacao) {
-		
-		Page<Transacao> transacoes = transacaoRepository.findAll(paginacao);
+	public Page<TransacaoDto> listar(Pageable paginacao, Usuario usuarioLogado) {
+
+		Page<Transacao> transacoes = transacaoRepository.findAllByUsuario(paginacao, usuarioLogado);
 		return transacoes.map(t -> modelMapper.map(t, TransacaoDto.class));
 
 	}
-	@Transactional
-	public TransacaoDto createTransacao(TransacaoFormDto transacaoFormDto) {
-		 try {
-		Transacao transacaoToSave = modelMapper.map(transacaoFormDto, Transacao.class);
-		transacaoToSave.setId(null);
-		
-		transacaoToSave.setUsuario(usuarioRepository.getById(transacaoFormDto.getUsuarioId()));
-    
-		
-		Transacao savedTransacao = transacaoRepository.save(transacaoToSave);
 
-		return modelMapper.map(savedTransacao, TransacaoDto.class);
-	     } catch (DataIntegrityViolationException e) {
-	            throw new DomainException("Usuario inválido");
-	        } catch (EntityNotFoundException e) {
-	            throw new ResourceNotFoundException("Usuario inválido");
-	        }
+	@Transactional
+	public TransacaoDto cadastrar(TransacaoFormDto transacaoFormDto, Usuario usuarioLogado) {
+
+		try {
+			Usuario usuario = usuarioRepository.getById(transacaoFormDto.getUsuarioId());
+
+			if (!usuario.equals(usuarioLogado)) {
+				throw obterExcecaoDeAcessoNegado();
+			}
+
+			Transacao transacaoToSave = modelMapper.map(transacaoFormDto, Transacao.class);
+			transacaoToSave.setId(null);
+
+			transacaoToSave.setUsuario(usuarioRepository.getById(transacaoFormDto.getUsuarioId()));
+
+			Transacao savedTransacao = transacaoRepository.save(transacaoToSave);
+
+			return modelMapper.map(savedTransacao, TransacaoDto.class);
+
+		} catch (DataIntegrityViolationException e) {
+			throw new DomainException("Usuario inválido");
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Usuario inválido");
+		}
 	}
 
 	@Transactional(readOnly = true)
-	public TransacaoDetalhadaDto mostrar(Long id) {
-		
-		var transacao = transacaoRepository.findById(id)
-				.orElseThrow(() -> new ResourceNotFoundException("Transacao não encontrada: " + id));
+	public TransacaoDetalhadaDto mostrar(Long id, Usuario usuarioLogado) {
 
+		Transacao transacao = transacaoRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Transacao não encontrada: " + id));
+		if (!transacao.getUsuario().equals(usuarioLogado)) {
+			throw obterExcecaoDeAcessoNegado();
+		}
 		return modelMapper.map(transacao, TransacaoDetalhadaDto.class);
-		
+
 	}
 
 	@Transactional
-	public TransacaoDto atualizar(TransacaoUpdateFormDto transacaoUpdateFormDto) {
+	public TransacaoDto atualizar(TransacaoUpdateFormDto transacaoUpdateFormDto, Usuario usuarioLogado) {
 		try {
-			var transacao = transacaoRepository.getById(transacaoUpdateFormDto.getId());
+			Transacao transacao = transacaoRepository.getById(transacaoUpdateFormDto.getId());
 
+			if (!transacao.getUsuario().equals(usuarioLogado)) {
+				throw obterExcecaoDeAcessoNegado();
+			}
 			transacao.atualizarInformacoes(transacaoUpdateFormDto.getTicker(), transacaoUpdateFormDto.getPreco(),
 					transacaoUpdateFormDto.getQuantidade(), transacaoUpdateFormDto.getTipo(),
 					transacaoUpdateFormDto.getData());
@@ -91,17 +102,27 @@ public class TransacaoService {
 	}
 
 	@Transactional
-	public void remover(Long id) {
+	public void remover(Long id, Usuario usuarioLogado) {
 		try {
-            transacaoRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Transacao inexistente: " + id);
-        }
-    }
-	 /*private MessageResponseDto createMessageResponse(Long id, String message) {
-	        return MessageResponseDto
-	                .builder()
-	                .message(message + id)
-	                .build();
-	    }*/
+
+			Transacao transacao = transacaoRepository.getById(id);
+			if (!transacao.getUsuario().equals(usuarioLogado)) {
+				throw obterExcecaoDeAcessoNegado();
+			}
+
+			transacaoRepository.deleteById(id);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Transacao inexistente: " + id);
+		}
+	}
+
+	private AccessDeniedException obterExcecaoDeAcessoNegado() {
+
+		return new AccessDeniedException("Acesso negado!");
+
+	}
+	/*
+	 * private MessageResponseDto createMessageResponse(Long id, String message) {
+	 * return MessageResponseDto .builder() .message(message + id) .build(); }
+	 */
 }
